tienda_virtual/requirements.txt
asgiref==3.7.2
boto3==1.28.17
botocore==1.31.17
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-dotenv==1.4.2
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jmespath==1.0.1
Pillow==10.0.0
psycopg2==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
tzdata==2023.3
urllib3==1.26.16
